feat: reworking infra, added the embed ECS task
